renamed config to aguments
