+++ mttl/models/library/retrievers.py
@@ -3,7 +3,7 @@ import copy
 import numpy as np
 import torch
 
-from mttl.config import ExpertConfig
+from mttl.arguments import ExpertConfig
 from mttl.logging import logger
 from mttl.models.expert_model import MultiExpertModel
 from mttl.models.library.expert import Expert
