+++ tests/conftest.py
@@ -9,7 +9,7 @@ import torch
 from transformers.models.llama.configuration_llama import LlamaConfig
 from transformers.models.llama.modeling_llama import LlamaForCausalLM
 
-from mttl.config import ExpertConfig, MoEExpertConfig, MultiExpertConfig
+from mttl.arguments import ExpertConfig, MoEExpertConfig, MultiExpertConfig
 from mttl.dataloader.flan_utils import download_flan
 from mttl.datamodule.mt_seq_to_seq_module import FlanConfig, FlanModule
 from mttl.models.expert_model import MultiExpertModel
@@ -186,7 +186,7 @@ def tmp_peer_moe_config(tmp_path: Path):
 
 @pytest.fixture
 def tmp_moe_exp_config(tmp_path):
-    from mttl.config import MoEExpertConfig
+    from mttl.arguments import MoEExpertConfig
 
     return MoEExpertConfig(
         model="EleutherAI/gpt-neo-125m",
