+++ tests/test_adapter_ranker.py
@@ -1,7 +1,7 @@
 # unit test for adapter_ranker
 import pytest
 
-from mttl.config import ExpertConfig, RankerConfig
+from mttl.arguments import ExpertConfig, RankerConfig
 from mttl.datamodule.mt_seq_to_seq_module import FlanConfig, FlanModule
 from mttl.models.containers.selectors.base import TaskPredictorSelector
 from mttl.models.expert_model import MultiExpertConfig, MultiExpertModel
